Implement unit context as class
